Move map builder to the package level
