+++ server/src/main/java/io/spine/server/event/enrich/EnrichmentsMap.java
@@ -20,30 +20,17 @@
 
 package io.spine.server.event.enrich;
 
-import com.google.common.base.Splitter;
-import com.google.common.collect.Collections2;
 import com.google.common.collect.ImmutableCollection;
 import com.google.common.collect.ImmutableMultimap;
-import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Message;
 import io.spine.Resources;
-import io.spine.option.OptionsProto;
-import io.spine.type.KnownTypes;
 import io.spine.type.TypeName;
-import io.spine.type.TypeUrl;
 
 import java.util.Collection;
-import java.util.List;
 import java.util.Properties;
 import java.util.Set;
-import java.util.regex.Pattern;
 
-import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Preconditions.checkNotNull;
-import static com.google.common.collect.Sets.newHashSet;
-import static com.google.protobuf.Descriptors.FieldDescriptor;
 import static io.spine.io.PropertyFiles.loadAllProperties;
-import static java.util.stream.Collectors.toList;
 
 /**
  * A map from an event enrichment Protobuf type name to the corresponding
@@ -55,20 +42,12 @@ import static java.util.stream.Collectors.toList;
  */
 final class EnrichmentsMap {
 
-    /** A separator between event types in the `.properties` file. */
-    private static final String EVENT_TYPE_SEPARATOR = ",";
-
-    private static final Splitter eventTypeSplitter = Splitter.on(EVENT_TYPE_SEPARATOR);
-
-    private static final Pattern pipeSeparatorPattern = Pattern.compile("\\|");
-
-    private static final char PROTO_PACKAGE_SEPARATOR = '.';
-
     /** A map from enrichment class name to enriched message class name. */
     private static final ImmutableMultimap<String, String> enrichmentsMap = buildEnrichmentsMap();
 
     /** Prevents instantiation of this utility class. */
-    private EnrichmentsMap() {}
+    private EnrichmentsMap() {
+    }
 
     /** Obtains immutable map from enrichment class name to enriched message class name. */
     static ImmutableMultimap<String, String> instance() {
@@ -84,143 +63,9 @@ final class EnrichmentsMap {
 
     private static ImmutableMultimap<String, String> buildEnrichmentsMap() {
         Set<Properties> propertiesSet = loadAllProperties(Resources.ENRICHMENTS);
-        Builder builder = new Builder(propertiesSet);
+        EnrichmentMapBuilder builder = new EnrichmentMapBuilder(propertiesSet);
         ImmutableMultimap<String, String> result = builder.build();
         return result;
     }
 
-    private static class Builder {
-
-        /**
-         * Constant indicating a package qualifier.
-         *
-         * <p>Must be a postfix of the qualifier.
-         */
-        private static final String PACKAGE_WILDCARD_INDICATOR = ".*";
-
-        private final Iterable<Properties> properties;
-        private final ImmutableMultimap.Builder<String, String> builder;
-
-        private Builder(Iterable<Properties> properties) {
-            this.properties = properties;
-            this.builder = ImmutableMultimap.builder();
-        }
-
-        private ImmutableMultimap<String, String> build() {
-            for (Properties props : this.properties) {
-                put(props);
-            }
-            return builder.build();
-        }
-
-        private void put(Properties props) {
-            Set<String> enrichmentTypes = props.stringPropertyNames();
-            for (String enrichmentType : enrichmentTypes) {
-                String eventTypesStr = props.getProperty(enrichmentType);
-                Iterable<String> eventTypes = eventTypeSplitter.split(eventTypesStr);
-                put(enrichmentType, eventTypes);
-            }
-        }
-
-        private void put(String enrichmentType, Iterable<String> eventQualifiers) {
-            for (String eventQualifier : eventQualifiers) {
-                if (isPackage(eventQualifier)) {
-                    putAllTypesFromPackage(enrichmentType, eventQualifier);
-                } else {
-                    builder.put(enrichmentType, eventQualifier);
-                }
-            }
-        }
-
-        /**
-         * Puts all the events from the given package into the map to match the
-         * given enrichment type.
-         *
-         * @param enrichmentType type of the enrichment for the given events
-         * @param eventsPackage  package qualifier representing the protobuf package containing
-         *                       the event to enrich
-         */
-        private void putAllTypesFromPackage(String enrichmentType, String eventsPackage) {
-            int lastSignificantCharPos = eventsPackage.length() -
-                    PACKAGE_WILDCARD_INDICATOR.length();
-            String packageName = eventsPackage.substring(0, lastSignificantCharPos);
-            Set<String> boundFields = getBoundFields(enrichmentType);
-            Collection<TypeUrl> eventTypes = KnownTypes.instance()
-                                                       .getAllFromPackage(packageName);
-            for (TypeUrl type : eventTypes) {
-                String typeQualifier = type.getTypeName();
-                if (hasOneOfTargetFields(typeQualifier, boundFields)) {
-                    builder.put(enrichmentType, typeQualifier);
-                }
-            }
-        }
-
-        private static Set<String> getBoundFields(String enrichmentType) {
-            Descriptor enrichmentDescriptor = TypeName.of(enrichmentType)
-                                                      .getMessageDescriptor();
-            Set<String> result = newHashSet();
-            for (FieldDescriptor field : enrichmentDescriptor.getFields()) {
-                String extension = field.getOptions()
-                                        .getExtension(OptionsProto.by);
-                Collection<String> fieldNames = parseFieldNames(extension);
-                result.addAll(fieldNames);
-            }
-            return result;
-        }
-
-        private static Collection<String> parseFieldNames(String qualifiers) {
-            Collection<String> result =
-                    pipeSeparatorPattern.splitAsStream(qualifiers)
-                                        .map(String::trim)
-                                        .filter(fieldName -> !fieldName.isEmpty())
-                                        .map(Builder::getSimpleFieldName)
-                                        .collect(toList());
-            return result;
-        }
-
-        @SuppressWarnings("ConstantConditions")
-        private static String getSimpleFieldName(String qualifier) {
-            int startIndex = qualifier.lastIndexOf(PROTO_PACKAGE_SEPARATOR) + 1;
-            startIndex = startIndex > 0 // 0 is an invalid value, see line above
-                         ? startIndex
-                         : 0;
-            String fieldName = qualifier.substring(startIndex);
-            return fieldName;
-        }
-
-        private static boolean hasOneOfTargetFields(String eventType,
-                                                    Collection<String> targetFields) {
-            Descriptor eventDescriptor = TypeName.of(eventType)
-                                                 .getMessageDescriptor();
-            List<FieldDescriptor> fields = eventDescriptor.getFields();
-            Collection<String> fieldNames = Collections2.transform(
-                    fields,
-                    input -> {
-                        checkNotNull(input);
-                        return input.getName();
-                    });
-            for (String field : targetFields) {
-                if (fieldNames.contains(field)) {
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        /**
-         * Returns {@code true} if the given qualifier is a package according to the contract
-         * of {@code "enrichment_for") option notation.
-         */
-        private static boolean isPackage(String qualifier) {
-            checkNotNull(qualifier);
-            checkArgument(!qualifier.isEmpty());
-
-            int indexOfWildcardChar = qualifier.indexOf(PACKAGE_WILDCARD_INDICATOR);
-            int qualifierLength = qualifier.length();
-
-            boolean result =
-                    indexOfWildcardChar == (qualifierLength - PACKAGE_WILDCARD_INDICATOR.length());
-            return result;
-        }
-    }
 }
